+++ release_notes.md
@@ -1,6 +1,6 @@
 ## New Features
-* Ability to pass `RetryOptions` object to `CallHttpAsync()` invocations to add & customize retry behavior
-* Adding `FailedRequestRetryOptions` to `DurableHttpRequest`; when using the `CallHttpAsync` overload that takes the full object, set this property in order to affect retry logic.
+* New `HttpRetryOptions` class for passing retry options to `CallHttpAsync()` invocations to add & customize retry behavior
+* Added `HttpRetryOptions` to `DurableHttpRequest`; when using the `CallHttpAsync` overload that takes the full object, set this property in order to affect retry logic.
 
 ## Bug fixes
 * Updated DurableTask.AzureStorage dependency to v1.9.2, which includes the following fixes:
@@ -9,4 +9,4 @@
 * * Fixed stuck orchestration with duplicate message warning issue
 
 ## Breaking Changes
-* `IDurableOrchestrationContext`'s `CallHttpAsync(HttpMethod, Uri, string)` overload now has a `SerializableRetryOptions` parameter which is identical to `RetryOptions` except without the `Handle` property.
+* `IDurableOrchestrationContext`'s `CallHttpAsync(HttpMethod, Uri, string)` overload now has a `HttpRetryOptions` parameter